+++ liquid/loaders.py
@@ -10,6 +10,8 @@ from .builtin.loaders import PackageLoader
 from .builtin.loaders import TemplateNamespace
 from .builtin.loaders import TemplateSource
 from .builtin.loaders import UpToDate
+from .builtin.loaders import make_choice_loader
+from .builtin.loaders import make_file_system_loader
 
 __all__ = (
     "BaseLoader",
@@ -19,6 +21,8 @@ __all__ = (
     "DictLoader",
     "FileExtensionLoader",
     "FileSystemLoader",
+    "make_choice_loader",
+    "make_file_system_loader",
     "PackageLoader",
     "TemplateNamespace",
     "TemplateSource",